requirements.txt
Django==2.1.4
django-filter==2.0.0
djangorestframework==3.9.0
djangorestframework-datatables==0.4.1
et-xmlfile==1.0.1
jdcal==1.4
Pillow==5.3.0
pytz
